Hack around to brute force match detection for now
Need to restructure to support the IntCode syntax described in the
problem statement. Also want to segregate file parsing from the input
modifications and move the matching logic outside of the CLI.

Additionally the CLI arg handling could use some rework. But ultimately
if multiple problems use this I'll look at turning it into a local lib
and the CLI bits won't matter so much.